comments from changelly

+++ mbw/src/test/java/com/mycelium/wallet/external/changelly/ExchangeLoggingServiceTest.java
@@ -7,6 +7,7 @@ import com.mycelium.wallet.external.changelly.model.Order;
 import org.junit.Test;
 
 import java.io.IOException;
+import java.math.BigDecimal;
 import java.text.SimpleDateFormat;
 import java.util.Date;
 
@@ -18,7 +19,7 @@ public class ExchangeLoggingServiceTest {
     @Test
     public void testLogging() {
         ChangellyTransactionOffer offer = new ChangellyTransactionOffer();
-        offer.amountTo = 111.222;
+        offer.amountTo = BigDecimal.valueOf(111.222);
         Order order = new Order();
         order.transactionId = "5b0f10a544b9dfd86595fcea39f822919ee425ec2c7a87a2a29c0467e2669769";
         order.exchangingAmount = "100.1111";
